requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.6
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
drf-yasg==1.17.0
firebase-admin==3.0.0
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==19.2
protobuf==3.10.0
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
